requirements.txt
Flask
gunicorn
python-dotenv
openai
supabase
bcrypt
pandas